cleanup(core): split performance report rendering into its own module
Move the presentation half (formatReport, recommendations, formatDuration, linkify, cache advice, the exit-payload builder, and the URL constants) into performance-report.ts. The lifecycle file now only collects timings and analyzes them; the report file only presents. One-way dependency (lifecycle -> report at runtime, report -> lifecycle for the PerformanceSummary type only). Added section banners to the lifecycle file for navigation.
